ismeta/backend/requirements.txt
# Core
Django>=5.0,<5.2
djangorestframework>=3.14,<4.0
django-cors-headers>=4.3
python-decouple>=3.8

# Database
psycopg[binary]>=3.1
django-filter>=23.5

# Async / queues
celery>=5.3
redis>=5.0
django-celery-beat>=2.5

# API docs
drf-spectacular>=0.27

# HTTP client
httpx>=0.27

# Auth
PyJWT>=2.8

# Validation
pydantic>=2.7

# Fuzzy matching
rapidfuzz>=3.9

# Excel
openpyxl>=3.1

# LLM SDKs
openai>=1.30
google-generativeai>=0.7
anthropic>=0.30

# Utilities
python-dateutil>=2.9

# Dev
pytest>=8.2
pytest-django>=4.8
pytest-asyncio>=0.23
pytest-cov>=5.0
vcrpy>=6.0
ruff>=0.4
mypy>=1.10
django-stubs>=5.0
djangorestframework-stubs>=3.14

# Security
bandit>=1.7
pip-audit>=2.7

# Observability (заложено с E1)
opentelemetry-api>=1.25
opentelemetry-sdk>=1.25
opentelemetry-instrumentation-django>=0.46b0
opentelemetry-instrumentation-celery>=0.46b0
opentelemetry-instrumentation-psycopg>=0.46b0
opentelemetry-instrumentation-requests>=0.46b0
opentelemetry-instrumentation-httpx>=0.46b0
opentelemetry-exporter-otlp>=1.25
structlog>=24.1   # структурированные логи с trace_id
